fixes from merge
