+++ docs-src/helpers/source_helpers.rb
@@ -22,7 +22,6 @@ module SourceHelpers
       }
     end
     file = files.find{|x| x =~ /#{name}\.js$/}
-    
     body = File.read(File.join(base, file))
     content = body.match(/\/\*.*(?=^\*\/)/m).to_s
     source = body.split(/\*\//, 2).last
@@ -33,7 +32,7 @@ module SourceHelpers
       name: file_name.gsub('-', '_').camelize
     }
     
-    if extended_class = source.match(/class\s+\w+\s+extends\s+(\w+)/).try(:[], 1)
+    if extended_class = source.match(/export\sdefault\sclass\s+\w+\s+extends\s+(\w+)/).try(:[], 1)
       if extended_class_path = source.match(/import[\s\{]+#{extended_class}[\s\}]+from\s+[\'\"]([\w\.\/\-]+)/).try(:[], 1)
         component_details[:extends] = extended_class_path.match(/(\w+).js$/).try(:[], 1)
         extended_component = component(component_details[:extends])
