+++ lib/komps/floater.js
@@ -304,8 +304,8 @@ export default class Floater extends KompElement {
                 if (this.onHide) this.onHide()
                 delete this._hideTimeout;
                 delete this._removing;
-                })
             })
+        }
         return this;
     }
