+++ lib/komps/input.js
@@ -131,7 +131,7 @@ Example
                 attribute: input.name.includes('[') ? input.name.replace(/\]$/, '').split(/\]?\[/) : input.name,
                 name: input.name,
                 options: Array.from(input.querySelectorAll('option')).map(x => x.value ? [x.value, x.innerText] : x.innerText),
-                value: input.value,
+                value: input.value == '' ? undefined : input.value,
                 multiple: input.multiple,
                 includeBlank: input.dataset.blank,
                 class: input.getAttribute('class'),
@@ -221,11 +221,8 @@ export default class Input {
         if (typeof this.load != "function") {
             this.load = v => v
         }
-        
         this.input = this.createInput(options);
-        if (options.value == undefined) {
-            this.input._loading = this._load();
-        }
+        this.input._loading = this._load();
         this.setupInputListener(this.inputChange.bind(this));
         this.setupRecordListener(this.recordChange.bind(this));
     }
